+++ dashboard/views.py
@@ -1,6 +1,7 @@
 import json
 import yaml
 
+from airone.lib.acl import ACLType
 from airone.lib.http import render
 from airone.lib.http import http_get, http_post
 from airone.lib.http import http_file_upload
@@ -24,6 +25,95 @@ IMPORT_INFOS = [
     {'model': 'Attribute', 'resource': AttrResource},
     {'model': 'AttributeValue', 'resource': AttrValueResource},
 ]
+
+
+def _search_entries(user, hint_entity_ids, hint_attrs, entry_name, hint_referral,
+                    entry_limit=CONFIG.MAXIMUM_SEARCH_RESULTS):
+    attr_names = [x['name'] for x in hint_attrs]
+    results = {'ret_count': 0, 'ret_values': []}
+    for hint_entity_id in hint_entity_ids:
+        entity = Entity.objects.filter(id=hint_entity_id, is_active=True).first()
+
+        # Check EntityAttr for which user does not have readable permission
+        non_permission_entityattrs = [
+            x.name for x in entity.attrs.filter(name__in=attr_names, is_active=True)
+            if not user.has_permission(x, ACLType.Readable)]
+
+        search_results = Entry.search_entries(
+            user,
+            [hint_entity_id],
+            hint_attrs,
+            entry_limit if len(results['ret_values']) < entry_limit else 0,
+            entry_name,
+            hint_referral=hint_referral),
+
+        results['ret_count'] += search_results[0]['ret_count']
+        for entryinfo in search_results[0]['ret_values']:
+
+            # Check number of result values
+            if len(results['ret_values']) >= CONFIG.MAXIMUM_SEARCH_RESULTS:
+                continue
+            # Check Entry for which user does not have readable permission
+            if (not entryinfo['permission']['is_public'] and
+                    entryinfo['permission']['default_permission'] < ACLType.Readable.id):
+                entry = Entry.objects.filter(id=entryinfo['entry']['id'], is_active=True).first()
+                if not entry:
+                    Logger.warning('Non exist entry (id:%s) is registered in ESS.' %
+                                   entryinfo['entry']['id'])
+                    continue
+                if not user.has_permission(entry, ACLType.Readable):
+                    ret_value = {
+                        'entity': entryinfo['entity'],
+                        'entry': entryinfo['entry'],
+                        'attrs': {},
+                        'permission': False,
+                    }
+                    if 'referrals' in entryinfo:
+                        ret_value['referrals'] = entryinfo['referrals']
+
+                    results['ret_values'].append(ret_value)
+                    continue
+
+            ret_value = {
+                'entity': entryinfo['entity'],
+                'entry': entryinfo['entry'],
+                'attrs': {},
+                'permission': True,
+            }
+            if 'referrals' in entryinfo:
+                ret_value['referrals'] = entryinfo['referrals']
+
+            # Check Attribute for which user does not have readable permission
+            for attr_name in attr_names:
+                if attr_name in non_permission_entityattrs:
+                    ret_value['attrs'][attr_name] = {'permission': False}
+                    continue
+
+                if attr_name not in entryinfo['attrs'].keys():
+                    continue
+
+                if (not entryinfo['attrs'][attr_name]['permission']['is_public'] and
+                        entryinfo['attrs'][attr_name]['permission']['default_permission'] <
+                        ACLType.Readable.id):
+                    entry = Entry.objects.filter(id=entryinfo['entry']['id'],
+                                                 is_active=True).first()
+                    if not entry:
+                        Logger.warning('Non exist entry (id:%s) is registered in ESS.' %
+                                       entryinfo['entry']['id'])
+                        continue
+                    attr = entry.attrs.filter(schema__name=attr_name, is_active=True).first()
+                    if not attr:
+                        continue
+                    if not user.has_permission(attr, ACLType.Readable):
+                        ret_value['attrs'][attr_name] = {'permission': False}
+                        continue
+
+                ret_value['attrs'][attr_name] = entryinfo['attrs'][attr_name]
+                ret_value['attrs'][attr_name]['permission'] = True
+
+            results['ret_values'].append(ret_value)
+
+    return results
 
 
 @airone_profile
@@ -151,52 +241,44 @@ def advanced_search_result(request):
     user = User.objects.get(id=request.user.id)
 
     recv_entity = request.GET.getlist('entity[]')
-    recv_attr = request.GET.getlist('attr[]')
     is_all_entities = request.GET.get('is_all_entities') == 'true'
-    has_referral = request.GET.get('has_referral') == 'true'
+    has_referral = request.GET.get('has_referral', False)
     attrinfo = request.GET.get('attrinfo')
     entry_name = request.GET.get('entry_name')
 
-    # check entity params
-    if not is_all_entities:
-        if not recv_entity:
-            return HttpResponse("The entity[] parameters are required", status=400)
-        if not all(
-                [Entity.objects.filter(id=x, is_active=True).exists() for x in recv_entity]):
-            return HttpResponse("Invalid entity ID is specified", status=400)
-
     # check attribute params
-    if not recv_attr and not attrinfo:
-        return HttpResponse("The attr[] or attrinfo parameters is required", status=400)
-
-    # build hint attrs from JSON encoded params,
-    # or attr[] the older param to keep backward compatibility
-    # TODO deprecate attr[]
-    hint_attrs = [{'name': x} for x in recv_attr]
-    if attrinfo:
-        try:
-            hint_attrs = json.loads(attrinfo)
-        except json.JSONDecodeError:
-            return HttpResponse("The attrinfo parameter is not JSON", status=400)
+    if not attrinfo:
+        return HttpResponse("The attrinfo parameters is required", status=400)
+    try:
+        hint_attrs = json.loads(attrinfo)
+    except json.JSONDecodeError:
+        return HttpResponse("The attrinfo parameter is not JSON", status=400)
     attr_names = [x['name'] for x in hint_attrs]
 
+    # check entity params
+    hint_entity_ids = []
     if is_all_entities:
         attrs = sum(
             [list(EntityAttr.objects.filter(name=x, is_active=True)) for x in attr_names], [])
-        entities = list(set([x.parent_entity.id for x in attrs if x]))
+        hint_entity_ids = list(set([x.parent_entity.id for x in attrs if x and
+                                   user.has_permission(x.parent_entity, ACLType.Readable)]))
     else:
-        entities = recv_entity
+        if not recv_entity:
+            return HttpResponse("The entity[] parameters are required", status=400)
+
+        for entity_id in recv_entity:
+            entity = Entity.objects.filter(id=entity_id, is_active=True).first()
+            if not entity:
+                return HttpResponse("Invalid entity ID is specified", status=400)
+
+            if user.has_permission(entity, ACLType.Readable):
+                hint_entity_ids.append(entity.id)
 
     return render(request, 'advanced_search_result.html', {
         'hint_attrs': hint_attrs,
-        'results': Entry.search_entries(user,
-                                        entities,
-                                        hint_attrs,
-                                        CONFIG.MAXIMUM_SEARCH_RESULTS,
-                                        entry_name,
-                                        hint_referral=has_referral),
+        'results': _search_entries(user, hint_entity_ids, hint_attrs, entry_name, has_referral),
         'max_num': CONFIG.MAXIMUM_SEARCH_RESULTS,
-        'entities': ','.join([str(x) for x in entities]),
+        'entities': ','.join([str(x) for x in hint_entity_ids]),
         'has_referral': has_referral,
         'is_all_entities': is_all_entities,
         'entry_name': entry_name,